Move files and make stuff internal
